+++ assets/js/jquery-fixedmenu/jquery-fixedmenu.js
@@ -92,7 +92,7 @@ function fixedMenu() {
 			 * Add .fixedmenu-fixed CSS class to menu
 			 * Show ghost div.
 			 */
-			var bottom = windowHeight - menu.height();
+			var bottom = windowHeight - menu.outerHeight();
 
 			var css = {
 				'position': 'fixed',